ai-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
pandas>=2.2.3
numpy>=2.2.0
scikit-learn>=1.6.0
statsmodels>=0.14.4
xgboost>=2.1.0
torch>=2.6.0
pmdarima>=2.0.4
psycopg2-binary>=2.9.10
redis>=5.2.0
python-dotenv>=1.0.1
httpx>=0.28.0
google-generativeai>=0.8.0
